优化代码

+++ common/src/main/java/com/eghm/common/enums/DataType.java
@@ -14,31 +14,31 @@ public enum DataType {
     /**
      * 个人权限
      */
-    SELF((byte)1, "本人数据权限"),
+    SELF(1, "本人数据权限"),
 
     /**
      * 本部门数据权限
      */
-    SELF_DEPT((byte)2, "本部门数据权限"),
+    SELF_DEPT(2, "本部门数据权限"),
 
     /**
      * 本部门及子部门数据权限
      */
-    DEPT((byte)3, "本部门及子部门数据权限"),
+    DEPT(3, "本部门及子部门数据权限"),
 
     /**
      * 自定义数据权限
      */
-    CUSTOM((byte)4, "自定义数据权限"),
+    CUSTOM(4, "自定义数据权限"),
 
     /**
      * 所有数据权限
      */
-    ALL((byte)5, "所有数据权限"),
+    ALL(5, "所有数据权限"),
 
     ;
 
-    private final byte value;
+    private final int value;
 
     private final String msg;
