+++ model/src/main/java/com/eghm/model/dto/feedback/FeedbackQueryRequest.java
@@ -18,7 +18,7 @@ public class FeedbackQueryRequest extends PagingQuery {
     private static final long serialVersionUID = 1054408796616316094L;
 
     @ApiModelProperty("分类信息(数据字典)")
-    private Byte classify;
+    private Integer classify;
 
     @ApiModelProperty("状态 false:待解决 true:已解决")
     private Boolean state;
