+++ model/src/main/java/com/eghm/model/dto/image/ImageAddRequest.java
@@ -1,6 +1,6 @@
 package com.eghm.model.dto.image;
 
-import com.eghm.model.annotation.Sign;
+import com.eghm.common.annotation.Sign;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
@@ -23,7 +23,7 @@ public class ImageAddRequest implements Serializable {
 
     @ApiModelProperty(value = "图片类型", required = true)
     @NotNull(message = "图片类型不能为空")
-    private Byte classify;
+    private Integer classify;
 
     @ApiModelProperty("备注信息")
     private String remark;
