+++ model/src/main/java/com/eghm/model/dto/operator/OperatorAddRequest.java
@@ -33,7 +33,7 @@ public class OperatorAddRequest implements Serializable {
     private String roleIds;
 
     @ApiModelProperty("数据权限类型")
-    private Byte permissionType;
+    private Integer permissionType;
 
     @ApiModelProperty("数据权限部门id,逗号分割")
     private String deptIds;
