+++ service/src/main/java/com/eghm/model/PushTemplate.java
@@ -34,7 +34,7 @@ public class PushTemplate {
     private String nid;
 
     @ApiModelProperty("状态 0:关闭 1:开启")
-    private Byte state;
+    private Integer state;
 
     @ApiModelProperty("消息内容")
     private String content;
