+++ service/src/main/java/com/eghm/model/SysNotice.java
@@ -19,23 +19,23 @@ public class SysNotice extends BaseEntity {
     /**
      * 未发布
      */
-    public static final byte STATE_0 = 0;
+    public static final int STATE_0 = 0;
 
     /**
      * 已发布
      */
-    public static final byte STATE_1 = 1;
+    public static final int STATE_1 = 1;
 
     @ApiModelProperty("公告标题")
     private String title;
 
     @ApiModelProperty("公告类型(数据字典表sys_notice_type)")
-    private Byte classify;
+    private Integer classify;
 
     @ApiModelProperty("公告内容(富文本)")
     private String content;
 
     @ApiModelProperty("是否发布 0:未发布 1:已发布")
-    private Byte state;
+    private Integer state;
 
 }
