+++ service/src/main/java/com/eghm/service/cache/CacheProxyService.java
@@ -35,7 +35,7 @@ public interface CacheProxyService {
      * @param classify banner所属模块,数据字典的值
      * @return 轮播图列表
      */
-    List<Banner> getBanner(Channel channel, Byte classify);
+    List<Banner> getBanner(Channel channel, Integer classify);
 
     /**
      * 新增轮播图信息
